requirements.txt

plotly>=6.0.1
vanna>=0.7.9
chromadb>=1.0.7
openai>=1.76.2
postgres>=4.0
python-slugify>=8.0.0
pillow>=10.0.0
# Adicionando nova versão do pacote requests
requests>=2.28.0
# Conexão com PostgreSQL
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.40
# Interface Web
streamlit>=1.45.0
# Utilitários
python-dotenv>=1.0.0
pandas>=2.2.3
# Processamento de dados
numpy>=2.2.5
# Exportação para Excel
xlsxwriter>=3.2.0
# Documentação
pydantic>=2.11.4
python-dateutil

# Dependências para desenvolvimento e testes
ipython>=8.12.0